Refactor file js and html to vue single file component
