--- form/components/relation/vue/relation.html
@@ -1,75 +1,0 @@
-<div style="margin-bottom: 5px;">
-  <div ref="relation_header_title" class="box-header with-border skin-color" style="width:100%; display: flex; font-weight: bold; font-size: 1.3em; align-items: center; margin-button:3px; background-color: #ffffff; ">
-    <span v-t="'plugins.editing.edit_relation'"></span>
-    <span style="margin-left: 2px;">: {{relation.name.toUpperCase()}}</span>
-  </div>
-  <div ref="relation_header_tools" class="box-header with-border" style="width:100%; display: flex; justify-content: flex-end; margin-button:3px; background-color: #ffffff; ">
-    <div id="search-box" style="margin-right: auto;">
-      <input v-if="relationsLength" type="text" class="form-control" id="filterRelation" :placeholder="placeholdersearch">
-    </div>
-    <div>
-      <span class="g3w-icon add-link" align="center"
-            v-if="capabilities.relation.find(capability => capability === 'change_attr_feature') !== undefined"
-            v-t-tooltip:bottom.create="tooltips.link_relation" @click="enableAddLinkButtons ? linkRelation() : null"
-            :class="[{'disabled': !enableAddLinkButtons}, g3wtemplate.font['link']]">
-      </span>
-      <span v-if="capabilities.relation.find(capability => capability === 'add_feature') !== undefined"
-            v-t-tooltip:bottom.create="tooltips.link_relation"
-            @click="enableAddLinkButtons ? addRelationAndLink() : null"
-            class="g3w-icon add-link pull-right"
-            :class="[{'disabled' : !enableAddLinkButtons}, g3wtemplate.font['plus']]">
-          </span>
-    </div>
-  </div>
-  <div ref="relation_body" class="box-body" style="padding:0;">
-    <template v-if="relationsLength">
-      <table class="table g3wform-relation-table table-striped" style="width:100%">
-        <thead>
-          <tr>
-            <th v-t="'tools'"></th>
-            <th></th>
-            <th v-for="attribute in relationAttributesSubset(relations[0])">
-              {{attribute.label}}
-            </th>
-          </tr>
-          </thead>
-          <tbody>
-            <tr v-for="(relation, index) in relations" class="featurebox-header">
-              <td>
-                <div style="display: flex">
-                  <div class="skin-tooltip-right editbtn enabled" @click="startTool(relationtool, index)"
-                       v-for="relationtool in getRelationTools()" :key="relationtool.state.name"
-                       data-toggle="tooltip"
-                       data-placement="right" v-t-tooltip:plugin="relationtool.state.name">
-                    <img height="20px" width="20px" :src="resourcesurl + 'images/'+ relationtool.state.icon"/>
-                  </div>
-                </div>
-              </td>
-              <td class="action-cell">
-                <div v-if="!fieldrequired && capabilities.relation.find(capability => capability === 'change_attr_feature') !== undefined"
-                     class="g3w-mini-relation-icon g3w-icon" :class="g3wtemplate.font['unlink']"
-                     @click="unlinkRelation(index)"
-                     v-t-tooltip:right.create="tooltips.unlink_relation"
-                     aria-hidden="true">
-                </div>
-              </td>
-              <td v-show="!showAllFieds(index)" v-for="attribute in relationAttributesSubset(relation)">
-                <template v-if="isMedia(attribute.value) && getValue(attribute.value)">
-                  <div class="preview">
-                    <a :href="getValue(attribute.value)" target="_blank">
-                      <div class="previewtype" :class="getMediaType(attribute.value.mime_type).type">
-                        <i class="fa-2x" :class="g3wtemplate.font[getMediaType(attribute.value.mime_type).type]"></i>
-                      </div>
-                    </a>
-                    <div class="filename">{{ getFileName(attribute.value) }}</div>
-                  </div>
-                </template>
-                <a v-else-if="isLink(attribute)" :href="getValue(attribute.value)" target="_blank">{{ getValue(attribute.value) }}</a>
-                <span v-else>{{ getValue(attribute.value) }}</span>
-              </td>
-            </tr>
-          </tbody>
-        </table>
-      </template>
-    </div>
-</div>
