--- form/components/relations/vue/relations.html
@@ -1,8 +1,0 @@
-<div id="feature_relations_form">
-  <relation
-    :relation="relation.relation"
-    :relations="relation.relations"
-    :resourcesurl="resourcesurl"
-    :formeventbus="formeventbus">
-  </relation>
-</div>
