--- table/vue/table.html
@@ -1,42 +1,0 @@
-<div id="editing_table" class="g3w-editing-table">
-  <div ref="editing_table_header" class="editing_table_header">
-    <div style="display: flex; justify-content: space-between; align-items: baseline">
-      <h3 class="editing_table_title">{{ state.title }}</h3>
-      <h-resize @h_size_change="changeSize"></h-resize>
-    </div>
-    <div class="editing_table_relation_messagge"  v-if="state.isrelation" style="font-weight: bold" v-t-plugin="'editing.relation.table.info'">
-    </div>
-  </div>
-  <table v-if="show" class="display" style="width:100%">
-    <thead>
-    <tr>
-      <th v-if="!state.isrelation" style="max-width: 60px"></th>
-      <th v-if="state.isrelation"></th>
-      <th v-for="header in state.headers">{{ header.label }}</th>
-    </tr>
-    </thead>
-    <tbody>
-    <tr v-for="(feature, index) in state.features" :key="feature.__gis3w_feature_uid" :id="feature.__gis3w_feature_uid">
-      <td v-if="!state.isrelation">
-        <div id="table-editing-tools" style="display:flex; justify-content: space-between;">
-          <i v-if="showTool('change_attr_feature')" :class="g3wtemplate.font['pencil']" class="g3w-icon skin-tooltip-right"  data-placement="right"  style="color:#30cce7;" v-t-title:plugin="'editing.table.edit'" aria-hidden="true" @click="editFeature(feature.__gis3w_feature_uid)"></i>
-          <i v-if="showTool('add_feature')" :class="g3wtemplate.font['copy-paste']" class="g3w-icon skin-tooltip-right"  data-placement="right"  style="color:#d98b14; padding: 5px 7px 5px 7px;" v-t-title:plugin="'editing.table.copy'" aria-hidden="true" @click="copyFeature(feature.__gis3w_feature_uid)"></i>
-          <i v-if="showTool('delete_feature')" :class="g3wtemplate.font['trash-o']" class="g3w-icon skin-tooltip-right"  data-placement="right"  style="color:red;" v-t-title:plugin="'editing.table.delete'" aria-hidden="true" @click="deleteFeature(feature.__gis3w_feature_uid)"></i>
-        </div>
-      </td>
-      <td v-if="state.isrelation">
-        <input :id="`relation__${index}`" @change="linkFeature(index, $event)" class="magic-checkbox" type="checkbox">
-        <label :for="`relation__${index}`"></label>
-      </td>
-      <td v-if ="showValue(key)" v-for="(value, key) in feature" :key="key">
-        <g3w-media v-if=" getValue(value) && isMediaField(key)" :state="value"></g3w-media>
-        <p v-else>{{ getValue(value) }}</p>
-      </td>
-    </tr>
-    </tbody>
-  </table>
-  <div id="buttons" ref="table_editing_footer_buttons" class="table_editing_footer_buttons" style="position: absolute; bottom: 10px; width: 100%; display:flex; justify-content: center;">
-    <button v-t-plugin="state.isrelation ? 'editing.form.buttons.save_and_back' : 'editing.form.buttons.save'" class="btn btn-success" style="margin-right: 10px" @click="save"></button>
-    <button v-t-plugin="'editing.form.buttons.cancel'" class="btn btn-danger" @click="cancel"></button>
-  </div>
-</div>
