--- vue/components/tool.html
@@ -1,3 +1,0 @@
-<div @click ="state.enabled ? toggletool(): null" v-if="state.visible" class="editbtn" :class="{'enabled' : state.enabled, 'toggled' : state.active}">
-  <img height="25px" width="25px" :alt="state.name" :src="resourcesurl + 'images/'+ state.icon" v-t-title:plugin="state.name"/>
-</div>
