--- vue/components/toolsoftool.html
@@ -1,14 +1,0 @@
-<div id="toolsoftoolcontainer" style="display: flex; flex-direction: column;
-    margin: 5px;
-    padding: 5px;
-    border-radius: 5px;">
-    <template v-for="tool in tools">
-      <component
-              :is="tool.type"
-              :options="tool.options">
-      </component>
-      <divider></divider>
-    </template>
-
-  </div>
-</div>
