--- vue/components/toolsoftool/tools.js
@@ -1,8 +1,0 @@
-import snap  from  './snap.vue';
-import measure from './measure.vue';
-const Tools = {
-  snap,
-  measure
-};
-
-module.exports = Tools;
