--- vue/editing.html
@@ -1,30 +1,0 @@
-<div class="g3w-editing-panel">
-  <div v-if="!appState.online" id="onlineofflinemessage" style="margin-bottom: 5px; padding: 5px; border-radius: 3px; background-color: orange; color:white; font-weight: bold">
-    <div v-t-plugin="'editing.messages.offline'"></div>
-  </div>
-  <div v-if="showcommitbar" style="display: flex; margin-bottom: 5px">
-    <div style="margin-right: auto;" class="editing-button" @click="canCommit ? commit(): null" :class="{'enabled' : canCommit }">
-      <span class="editing-icon" :class="g3wtemplate.font['save']"></span>
-    </div>
-    <div class="editing-button " @click="canUndo ? undo(): null" :class="{'enabled' : canUndo }">
-      <span class="editing-icon" :class="g3wtemplate.font['arrow-left']"></span>
-    </div>
-    <div class="editing-button "  @click="canRedo ? redo(): null" :class="{'enabled' : canRedo }">
-      <span class="editing-icon" :class="g3wtemplate.font['arrow-right']"></span>
-    </div>
-  </div>
-  <div v-else style="height: 10px;"></div>
-  <div id="toolboxes">
-    <toolbox
-      :state="toolboxstate"
-      :resourcesurl="resourcesurl"
-      @setselectedtoolbox="setSelectedToolbox"
-      @starttoolbox="startToolBox"
-      @stoptoolbox="stopToolBox"
-      @savetoolbox="saveToolBox"
-      @setactivetool="startActiveTool"
-      @stopactivetool="stopActiveTool"
-      v-for="toolboxstate in state.toolboxes" :key="toolboxstate.id">
-    </toolbox>
-  </div>
-</div>
