+++ workflows/steps/tasks/mergefeaturestask.js
@@ -1,9 +1,8 @@
-const inherit = g3wsdk.core.utils.inherit;
-const base =  g3wsdk.core.utils.base;
+const { base, inherit } = g3wsdk.core.utils;
 const { dissolve } = g3wsdk.core.geoutils;
 const EditingTask = require('./editingtask');
-const GUI = g3wsdk.gui.GUI;
-const SelectFeaturesDom = require('../../../vue/components/selectfeatures/selectfeatures');
+const { GUI } = g3wsdk.gui;
+const SelectFeaturesDom = require('../../../g3w-editing-components/selectfeatures');
 
 function MergeFeaturesTask(options={}){
   base(this, options);
